修复推荐算法的一些问题

+++ recommend-service/src/main/java/org/L2/recommend/domain/model/UserPreference.java
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 public class UserPreference {
     private Long userId;
     private String vectorJson;
-    private Integer playCount;
+    private Double playCount;
     private LocalDateTime snapshotAt;
 }
